+++ electrostoreAPI/Controllers/UserCommandController.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Authorization;
 using electrostore.Dto;
 using electrostore.Services.CommandCommentaireService;
+using System.Security.Claims;
 
 namespace electrostore.Controllers
 {
@@ -20,6 +20,10 @@ namespace electrostore.Controllers
         [HttpGet]
         public async Task<ActionResult<IEnumerable<ReadCommandCommentaireDto>>> GetCommandsCommentairesByUserId([FromRoute] int id_user, [FromQuery] int limit = 100, [FromQuery] int offset = 0)
         {
+            if (!User.IsInRole("Admin") && id_user != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? ""))
+            {
+                return Unauthorized(new { message = "You are not allowed to access this resource" });
+            }
             var commandCommentaires = await _commandCommentaireService.GetCommandsCommentairesByUserId(id_user, limit, offset);
             if (commandCommentaires.Result is BadRequestObjectResult)
             {
@@ -35,6 +39,10 @@ namespace electrostore.Controllers
         [HttpGet("{id_commandcommentaire}")]
         public async Task<ActionResult<ReadCommandCommentaireDto>> GetCommandsCommentaireById([FromRoute] int id_user, [FromRoute] int id_commandcommentaire)
         {
+            if (!User.IsInRole("Admin") && id_user != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? ""))
+            {
+                return Unauthorized(new { message = "You are not allowed to access this resource" });
+            }
             var commandCommentaire = await _commandCommentaireService.GetCommandsCommentaireById(id_commandcommentaire, id_user);
             if (commandCommentaire.Result is BadRequestObjectResult)
             {
@@ -52,6 +60,10 @@ namespace electrostore.Controllers
         [HttpPut("{id_commandcommentaire}")]
         public async Task<ActionResult<ReadCommandCommentaireDto>> UpdateCommandCommentaire([FromRoute] int id_user, [FromRoute] int id_commandcommentaire, [FromBody] UpdateCommandCommentaireDto commandCommentaireDto)
         {
+            if (!User.IsInRole("Admin") && id_user != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? ""))
+            {
+                return Unauthorized(new { message = "You are not allowed to access this resource" });
+            }
             var commandCommentaire = await _commandCommentaireService.UpdateCommentaire(id_commandcommentaire, commandCommentaireDto, id_user);
             if (commandCommentaire.Result is BadRequestObjectResult)
             {
@@ -67,6 +79,10 @@ namespace electrostore.Controllers
         [HttpDelete("{id_commandcommentaire}")]
         public async Task<ActionResult> DeleteCommandCommentaire([FromRoute] int id_user, [FromRoute] int id_commandcommentaire)
         {
+            if (!User.IsInRole("Admin") && id_user != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? ""))
+            {
+                return Unauthorized(new { message = "You are not allowed to access this resource" });
+            }
             await _commandCommentaireService.DeleteCommentaire(id_commandcommentaire, id_user);
             return NoContent();
         }
