+++ electrostoreAPI/Controllers/UserController.cs
@@ -2,7 +2,8 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Authorization;
 using electrostore.Dto;
 using electrostore.Services.UserService;
-using OneOf.Types;
+using electrostore.Services.JwtService;
+using System.Security.Claims;
 
 namespace electrostore.Controllers
 {
@@ -12,10 +13,12 @@ namespace electrostore.Controllers
     public class UserController : ControllerBase
     {
         private readonly IUserService _userService;
+        private readonly JwtService _jwtService;
 
-        public UserController(IUserService userService)
+        public UserController(IUserService userService, JwtService jwtService)
         {
             _userService = userService;
+            _jwtService = jwtService;
         }
 
         [HttpGet]
@@ -41,8 +44,19 @@ namespace electrostore.Controllers
         }
 
         [HttpPost]
+        [AllowAnonymous]
         public async Task<ActionResult<ReadUserDto>> CreateUser([FromBody] CreateUserDto userDto)
         {
+            //if a bearer token is provided, the user is already authenticated and if the user is an admin, he can create a user with any role
+            // if the user is not an admin, he can only create a user with the role "user"
+            if (User != null)
+            {
+                if (!User.IsInRole("Admin") && userDto.role_user != "user")
+                {
+                    return Unauthorized(new { message = "You are not allowed to create a user with this role" });
+                }
+            }
+
             var user = await _userService.CreateUser(userDto);
             if (user.Result is BadRequestObjectResult)
             {
@@ -58,6 +72,11 @@ namespace electrostore.Controllers
         [HttpPut("{id_user}")]
         public async Task<ActionResult<ReadUserDto>> UpdateUser([FromRoute] int id_user, [FromBody] UpdateUserDto userDto)
         {
+            if (!User.IsInRole("Admin") && id_user != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? ""))
+            {
+                return Unauthorized(new { message = "You are not allowed to update this user" });
+            }
+
             var user = await _userService.UpdateUser(id_user, userDto);
             if (user.Result is BadRequestObjectResult)
             {
@@ -73,12 +92,70 @@ namespace electrostore.Controllers
         [HttpDelete("{id_user}")]
         public async Task<ActionResult> DeleteUser([FromRoute] int id_user)
         {
+            if (!User.IsInRole("Admin") && id_user != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? ""))
+            {
+                return Unauthorized(new { message = "You are not allowed to delete this user" });
+            }
             var result = await _userService.DeleteUser(id_user);
             if (result is BadRequestObjectResult)
             {
                 return result;
             }
             return NoContent();
+        }
+    
+        [HttpPost("login")]
+        [AllowAnonymous]
+        public async Task<IActionResult> Login([FromBody] LoginRequest loginRequest)
+        {
+            var user = await _userService.GetUserByEmail(loginRequest.Email);
+            if (user.Result is BadRequestObjectResult)
+            {
+                return Unauthorized(new { message = "Invalid credentials" });
+            }
+            if (!await _userService.CheckUserPassword(loginRequest.Email, loginRequest.Password))
+            {
+                return Unauthorized(new { message = "Invalid credentials" });
+            }
+            if (user.Value == null)
+            {
+                return StatusCode(500);
+            }
+
+            var token = _jwtService.GenerateToken(user.Value);
+            return Ok(new { Token = token });
+        }
+
+        [HttpPost("forgot-password")]
+        [AllowAnonymous]
+        public async Task<IActionResult> ForgotPassword([FromBody] ForgotPasswordRequest forgotPasswordRequest)
+        {
+            var result = await _userService.ForgotPassword(forgotPasswordRequest);
+            if (result is BadRequestObjectResult)
+            {
+                return result;
+            }
+            if (result == null)
+            {
+                return StatusCode(500);
+            }
+            return Ok();
+        }
+
+        [HttpPost("reset-password")]
+        [AllowAnonymous]
+        public async Task<IActionResult> ResetPassword([FromBody] ResetPasswordRequest resetPasswordRequest)
+        {
+            var result = await _userService.ResetPassword(resetPasswordRequest);
+            if (result is BadRequestObjectResult)
+            {
+                return result;
+            }
+            if (result == null)
+            {
+                return StatusCode(500);
+            }
+            return Ok();
         }
     }
 }
