+++ electrostoreAPI/Controllers/UserProjetController.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Authorization;
 using electrostore.Dto;
 using electrostore.Services.ProjetCommentaireService;
+using System.Security.Claims;
 
 namespace electrostore.Controllers
 {
@@ -20,6 +20,10 @@ namespace electrostore.Controllers
         [HttpGet]
         public async Task<ActionResult<IEnumerable<ReadProjetCommentaireDto>>> GetProjetCommentairesByUserId([FromRoute] int id_user, [FromQuery] int limit = 100, [FromQuery] int offset = 0)
         {
+            if (!User.IsInRole("Admin") && id_user != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? ""))
+            {
+                return Unauthorized(new { message = "You are not allowed to access this resource" });
+            }
             var projetCommentaires = await _projetCommentaireService.GetProjetCommentairesByUserId(id_user, limit, offset);
             if (projetCommentaires.Result is BadRequestObjectResult)
             {
@@ -35,6 +39,10 @@ namespace electrostore.Controllers
         [HttpGet("{id_projetcommentaire}")]
         public async Task<ActionResult<ReadProjetCommentaireDto>> GetProjetCommentairesByCommentaireId([FromRoute] int id_user, [FromRoute] int id_projetcommentaire)
         {
+            if (!User.IsInRole("Admin") && id_user != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? ""))
+            {
+                return Unauthorized(new { message = "You are not allowed to access this resource" });
+            }
             var projetCommentaire = await _projetCommentaireService.GetProjetCommentairesByCommentaireId(id_projetcommentaire, id_user);
             if (projetCommentaire.Result is BadRequestObjectResult)
             {
@@ -52,6 +60,10 @@ namespace electrostore.Controllers
         [HttpPut("{id_projetcommentaire}")]
         public async Task<ActionResult<ReadProjetCommentaireDto>> UpdateProjetCommentaire([FromRoute] int id_user, [FromRoute] int id_projetcommentaire, [FromBody] UpdateProjetCommentaireDto projetCommentaireDto)
         {
+            if (!User.IsInRole("Admin") && id_user != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? ""))
+            {
+                return Unauthorized(new { message = "You are not allowed to access this resource" });
+            }
             var projetCommentaire = await _projetCommentaireService.UpdateProjetCommentaire(id_projetcommentaire, projetCommentaireDto, id_user);
             if (projetCommentaire.Result is BadRequestObjectResult)
             {
@@ -67,6 +79,10 @@ namespace electrostore.Controllers
         [HttpDelete("{id_projetcommentaire}")]
         public async Task<ActionResult> DeleteProjetCommentaire([FromRoute] int id_user, [FromRoute] int id_projetcommentaire)
         {
+            if (!User.IsInRole("Admin") && id_user != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? ""))
+            {
+                return Unauthorized(new { message = "You are not allowed to access this resource" });
+            }
             await _projetCommentaireService.DeleteProjetCommentaire(id_projetcommentaire, id_user);
             return NoContent();
         }
